+++ README.md
@@ -1,2 +1,3 @@
 # MLoL
+
 An application that provides suggestions on which league of legends champion is the most suitable for your playstyle.
